Felles 4210, fjerne ubrukte lenker, legg til los/risk
