+++ pom.xml
@@ -35,7 +35,7 @@
 
         <fp-uttak.version>14.3.1</fp-uttak.version>
         <svp-uttak.version>2.3.4</svp-uttak.version>
-        <felles.version>4.2.8</felles.version>
+        <felles.version>4.2.10</felles.version>
         <prosesstask.version>3.1.13</prosesstask.version>
         <fp-jms.version>2.2.0</fp-jms.version>
