--- web/src/main/java/no/nav/foreldrepenger/web/app/tjenester/behandling/vilkår/VilkårRestTjeneste.java
@@ -1,76 +1,0 @@
-package no.nav.foreldrepenger.web.app.tjenester.behandling.vilkår;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.transaction.Transactional;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.CacheControl;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import no.nav.foreldrepenger.behandlingslager.behandling.Behandlingsresultat;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingsresultatRepository;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
-import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
-import no.nav.foreldrepenger.web.app.tjenester.behandling.dto.BehandlingAbacSuppliers;
-import no.nav.foreldrepenger.web.app.tjenester.behandling.dto.UuidDto;
-import no.nav.vedtak.sikkerhet.abac.BeskyttetRessurs;
-import no.nav.vedtak.sikkerhet.abac.TilpassetAbacAttributt;
-import no.nav.vedtak.sikkerhet.abac.beskyttet.ActionType;
-import no.nav.vedtak.sikkerhet.abac.beskyttet.ResourceType;
-
-@Path(VilkårRestTjeneste.BASE_PATH)
-@Produces(MediaType.APPLICATION_JSON)
-@ApplicationScoped
-@Transactional
-public class VilkårRestTjeneste {
-
-    static final String BASE_PATH = "/behandling";
-    private static final String VILKÅR_V2_PART_PATH = "/vilkar-v2";
-    public static final String VILKÅR_V2_PATH = BASE_PATH + VILKÅR_V2_PART_PATH;
-
-    private BehandlingRepository behandlingRepository;
-    private BehandlingsresultatRepository behandlingsresultatRepository;
-
-    public VilkårRestTjeneste() {
-        // for CDI proxy
-    }
-
-    @Inject
-    public VilkårRestTjeneste(BehandlingRepositoryProvider behandlingRepositoryProvider) {
-        this.behandlingRepository = behandlingRepositoryProvider.getBehandlingRepository();
-        this.behandlingsresultatRepository = behandlingRepositoryProvider.getBehandlingsresultatRepository();
-    }
-
-    @GET
-    @Path(VILKÅR_V2_PART_PATH)
-    @Operation(description = "Hent informasjon om vilkår for en behandling", tags = "vilkår", responses = {
-            @ApiResponse(responseCode = "200", description = "Returnerer vilkår på behandling, tom liste hvis ikke eksisterer (GUI støtter ikke NOT_FOUND p.t.)", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = VilkårDto.class)))
-    })
-    @BeskyttetRessurs(actionType = ActionType.READ, resourceType = ResourceType.FAGSAK)
-    public Response getVilkår(@TilpassetAbacAttributt(supplierClass = BehandlingAbacSuppliers.UuidAbacDataSupplier.class)
-        @NotNull @QueryParam(UuidDto.NAME) @Parameter(description = UuidDto.DESC) @Valid UuidDto uuidDto) {
-        var behandling = behandlingRepository.hentBehandling(uuidDto.getBehandlingUuid());
-
-        var dto = VilkårDtoMapper.lagVilkarDto(behandling, getBehandlingsresultat(behandling.getId()));
-        var cc = new CacheControl();
-        cc.setNoCache(true);
-        cc.setNoStore(true);
-        cc.setMaxAge(0);
-        return Response.ok(dto).cacheControl(cc).build();
-    }
-
-    private Behandlingsresultat getBehandlingsresultat(Long behandlingId) {
-        return behandlingsresultatRepository.hentHvisEksisterer(behandlingId).orElse(null);
-    }
-}
